Moving the enums query test into another query integration test, to reuse the existing query testing infrastructure
