+++ blueflood-core/src/integration-test/java/com/rackspacecloud/blueflood/io/AstyanaxReaderIntegrationTest.java
@@ -91,7 +91,7 @@ public class AstyanaxReaderIntegrationTest extends IntegrationTestBase {
     @Test
     public void testCanReadEnumValuesByOneLocator() throws Exception {
         List<Locator> locatorList = new ArrayList<Locator>();
-        IMetric metric = writeEnumMetric("enum_metric1");
+        IMetric metric = writeEnumMetric("enum_metric1","333333");
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.TYPE.name().toLowerCase(), null);
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.ROLLUP_TYPE.name().toLowerCase(), RollupType.ENUM.toString());
         locatorList.add(metric.getLocator());
@@ -117,12 +117,12 @@ public class AstyanaxReaderIntegrationTest extends IntegrationTestBase {
     @Test
     public void testCanReadEnumValuesByMultipleLocators() throws Exception {
         List<Locator> locatorList = new ArrayList<Locator>();
-        IMetric metric = writeEnumMetric("enum_metric2");
+        IMetric metric = writeEnumMetric("enum_metric2","333333");
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.TYPE.name().toLowerCase(), null);
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.ROLLUP_TYPE.name().toLowerCase(), RollupType.ENUM.toString());
         locatorList.add(metric.getLocator());
 
-        IMetric metric1 = writeEnumMetric("enum_metric3");
+        IMetric metric1 = writeEnumMetric("enum_metric3","333333");
         MetadataCache.getInstance().put(metric1.getLocator(), MetricMetadata.TYPE.name().toLowerCase(), null);
         MetadataCache.getInstance().put(metric1.getLocator(), MetricMetadata.ROLLUP_TYPE.name().toLowerCase(), RollupType.ENUM.toString());
         locatorList.add(metric1.getLocator());
@@ -152,13 +152,13 @@ public class AstyanaxReaderIntegrationTest extends IntegrationTestBase {
     public void testCanReadMixedEnumAndRegularMetrics() throws Exception {
         Set<Locator> enumLocatorsSet = new HashSet<Locator>();
         List<Locator> locatorList = new ArrayList<Locator>();
-        IMetric metric = writeEnumMetric("enum_metric3");
+        IMetric metric = writeEnumMetric("enum_metric3","333333");
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.TYPE.name().toLowerCase(), null);
         MetadataCache.getInstance().put(metric.getLocator(), MetricMetadata.ROLLUP_TYPE.name().toLowerCase(), RollupType.ENUM.toString());
         locatorList.add(metric.getLocator());
         enumLocatorsSet.add(metric.getLocator());
 
-        IMetric metric1 = writeEnumMetric("enum_metric4");
+        IMetric metric1 = writeEnumMetric("enum_metric4","333333");
         MetadataCache.getInstance().put(metric1.getLocator(), MetricMetadata.TYPE.name().toLowerCase(), null);
         MetadataCache.getInstance().put(metric1.getLocator(), MetricMetadata.ROLLUP_TYPE.name().toLowerCase(), RollupType.ENUM.toString());
         locatorList.add(metric1.getLocator());
